Moves bookarks manager from js to app
Resolves #10075

Auditors: @bsclifton

Test Plan:
- check if bookmarks manager is displayed correctly
